Add welcome message and default language pair
If a campaign ID is set, show a welcome message. Also remove the default
campaign ID of 'article-recommender-1' so as not to show the welcome
message when no campaign ID is set explicitly.

If the campaign ID is "WikiGapFinder" set the default source language to
"English" and the target language to "Swedish".

Bug: T244587
Change-Id: I84134586f6e7ff153e747408e720ad384db790de