Update labs_setup.sh
These are the changes I had to make to the setup script in order to move
to a larger labs instance.

Make sure depedency wheels are generated on the same machine to avoid
version mismatch. Also set it up so that the `related_articles` service
is disabled by default (it has its own labs instance).

Bug: T244587
Change-Id: I425d679796a0b9b8efae105adb853c0cb1755f2c