Load preprocessed numpy arrays from swift
We run memory usage tests in T339890#9155985, T339890#9156601,
T339890#9159257, and T339890#9162780. The 4th test where float
was downcasted to np.float32 and numpy arrays preprocessed,
resulted in a 5x drop in memory usage. This was the best
result of the 4 tests performed.

In this patch we are adapting the rec-api to fetch and use
the float32 preprocessed numpy arrays from swift.

Bug: T346218
Change-Id: I0f88b3b0b0e7abca58f868c2ca313be934591f74