Update repo docs to match LiftWing API GW docs
Aligned the old docs in this repo with the same information on
the LiftWing API gateway docs. This update resolves conflicting
details to ensure users have matching information as they migrate
from the rec-api instance on wmflabs to the new instance on LiftWing.

Also added a deprecation note that points users to the new resources.

Bug: T308164
Change-Id: Id48f993f4a4ac27e485f74e8702733e8a1d56585