API Continue support
Add a `fetch_all` argument to get() that triggers
the continue logic and make it return a list of
responses instead of a single response.

Continue support must be enabled (byusing the fetch_all
argument) and handled (by looping on the responses) on
a case-by-case basis. In most get requests, like the
various searches, we don't want continue support by default.

Bug: T379037
Change-Id: Iba0972e58b9eabfcba84340046ccf4e6c8296683