Fix the broken /docs url in production
The request for /openapi.json was not being redirected to the
actual path /service/lw/recommendation/openapi.json in production.

Because of that https://api.wikimedia.org/service/lw/recommendation/docs is
failing to render the docs.

Pass the API_PREFIX to FastAPI to fix the issue.

See https://fastapi.tiangolo.com/advanced/behind-a-proxy/#providing-the-root_path
for documentation.

API_PREFIX is an environment variable and should be overridden in production.

API_PREFIX=/service/lw/recommendation/api/

Additionally pass the servers option to target staging and production
from same openapi interface.

Bug: T371465
Change-Id: Ica39b94af4c8ae90afd800e381813496bb14327b