Add cache for translation campaigns pages
The cache key includes the revision id. This way, the cache is invalidated
when the revision is updated.

Cache will have the campaign definition(yet to be implemented, for now page metadata is used) and the links in page. For each link in page, all
interwiki information is included.

The caching technology is diskcache, but in future it need to be
replaced by a redis cache for better persistance and performance.

Bug: T371515
Change-Id: I4ccaaca27efc9af82b54fb984991d56b875e025b