Remove parameter from WIKIMEDIA_API url
Currently, the WIKIMEDIA_API configuration parameter which is used to
fetch the campaign pages, expects a "source" parameter to be given,
so that the API URL is properly constructed. This approach is similar
to the approach with use with Wikipedia APIs, but here it has no value,
because we don't hit different API for different languages, but rather
we always send requests to the same endpoint (the MetaWiki API).

Since, there is no need for this "source" parameter, this patch removes
it.

Change-Id: I3c6c9e8668df6c8be92ddb002c4daa92116d2dea