Add concurrency and error handling for fetch section suggestion requests
For fetching section suggestions for a list of candidates, we send
several parallel requests to CXServer API. We wait until the first
N requests are completed (where N, the desired number of results),
and we cancel the rest of the requests.

However, if a request to CXServer fails, it leads to an "Internal
server error". This patch adds error handling for such failures,
so that failed requests to CXServer are skipped, and don't break
the workflow.

Additionally, this patch adds a new configuration parameter
CXSERVER_API_CONCURRENCY to control the number of parallel
requests to CXServer.

Bug: T374387
Change-Id: I882eb99dcb747f58abfb0b44d20457584a9f1d66