Initialize campaign cache and update it every 1 hour
This patch follows a very simple, yet effective, approach to support
serving campaign recommendations from cache. When application is starting
up, the cache is initialized with the campaigns information and their
respective articles. Depending on how many campaigns are there, this can
take a while. This is done to avoid the delay when the first
recommendation request is made.

After the initialization, campaigns and their articles are only fetched
from cache, and no HTTP requests are being sent to other APIs, for
campaign recommendation requests. This means, any new campaign page
created will be picked up by the cache only after the cache is updated.
Also, any updates to the campaign definition or the list of articles in
the campaign will be picked up only after the cache is updated.

An asynchronous job is registered to update the cache every hour.

The patch is based on Santhosh's patches:
I1172ab6715b82e0238b12c5d36496abe3ba706ec, and
I480064cdded6095bfe22c985d1ff8a7123f74fb4, but simplifies the
caching/restoration process for campaigns.

Changelog:

- get_campaign_candidates
  * used to call meta API to fetch campaign pages. Now we just load all
    campaign-related+articles info from cache.
  * we do not update the cache anymore during HTTP request handling. We
    do that in a cronjob
- models.py
  * add __hash__ for the models to facilitate set operations
  * use sets instead of lists for article and candidates so that we
    get automatic deduplication, inside TranslationCampaign model
  * add CampaignMetadata model to encapsulate the required campaign
    metadata for each campaign. Used as a value object
  * add TranslationCampaignCollection to model a list of translation
   campaigns. Convenient for saving/restoring all campaigns to cache,
   with validation.
- cache.py
  * replace "set/get_campaign_page" with "set/get_translation_campaigns",
    as we'll only cache a single list containing all campaigns and their
    articles. We do not cache each campaign separately anymore.
- cache_updater.py
  * new file to support cache init/update functionality
  * uses a stub "get_campaign_metadata_by_pages" to populate the
    campaigns with dummy metadata. This will be replaced by an actual
    call to the "pagecollectionsmetadata" API endpoint in following
    patches.

Change-Id: I987f00d844465a580da45bdd3a07c9b9330c82b4