Use category search to find campaign pages instead of template
In order to fetch the campaign pages, we send a request to the
MetaWiki Action API. This request searches for pages that contain
a "Translation_campaign" template. However, this search is based
on an old assumption.

The new approach (I5f4b0041a4d3281702f2005ee1485b49610f8554) uses
an HTML marker, and a parser hook adds a tracking category to the
pages that contain that specific HTML marker.

To support fetching campaign pages based on the new approach, this
patch sends a query request to the MetaWiki Action API to fetch
pages that contain the specific tracking category that is added to
the pages that contain the campaign HTML marker.

Bug: T373132
Depends-On: I5f4b0041a4d3281702f2005ee1485b49610f8554
Change-Id: Ia08f27b7a139c93bc113e1c60e80f9c7c0fb225c