Filter out wikidata articles that don't have language links
It should be pretty rare that page collections include
links to wikidata item that don't have langlinks but
QID aren't very expressive (Q123456789) so typos are
easy to make and hard to spot.

Those items would never be served as recommendations
so might as well not store them in the cache.

Change-Id: I7f3112ab843dbfa323ef06acc5e04a24614126e1