Use random sorting only for topic-based recommendations
"random" sorting is used to diversify the topic-based recommendations
and avoid returning recommendations dominated by narrowly focused
topics. However, this is not the case when seeds are used, as we would like to get results as relevant as possible to the given seed. Since
"relevance" is the default for "gsrsort" parameter, we can just omit it
for these requests.

Bug: T377124
Change-Id: I87a1b3e694a2cf0c5c44882bb4d2cf43cbcab799