update cache in a single thread
This patch checks the worker id using the worker process id, and runs
the cache initialization/update only when the
"pid modulo number_of_workers" is equal to 0. Given that worker pids
are consecutive numbers, only one of them will have the above value
equal to 0.

Right now, the number_of_workers is hardcoded to 4, which is the case
for development environment, but this may be different for production.
Still, this is an improvement also for production, as it will
drastically reduce the number of workers to will run the cache update
process.

To properly, fix the issue for production we should set the number of
workers as an env variable and use that instead of the hardcoded value.

Bug: T379036
Change-Id: I6a0af171a866af67cc5d1469571a63c6b54dbca7