Avoid following redirects in external API calls, improve error handling
If redirect is followed, the external API call will be made to the
redirected URL without going through the proxy URL and header setup
we have in production.

The wikidata requests were always redirecting with 301 due to two issues
1. The API URL must be https://www.wikidata.org/w/api.php and not https://wikidata.org/w/api.php
2. The Host header value must be www.wikidata.org and not wikidata.org

If any of the above issues are present, there will be a 301 redirect and
redirect url will use https://www.wikidata.org instead of localhost:6500
setup in production.

To fix this issue
1. Correct the configuration with www prefix in URL and header
2. Disable redirect following in external get requests. Avoid surprises.

Additionally, improve logging and exception handling
* Add Host headers to the configuration by default. They are same even
  if API url is via proxy or not.
* In the external get API calls, assert host header is present. If not,
  raise an exception. This is to catch any misconfiguration in production
* Improve error logging and http request logging. Print all external API
  URL and header in debug logs

To test, try changing the WIKIPEDIA_API_HEADER value with some typos and
observer exceptions in the log. Or try removing www from wikidata URLs

Bug: T379592
Change-Id: I4220eb6e5ebb0c3396cfdafda0673143d8a4613d