performance: Process only the required number of collection articles
Currently, all items in all collections are processed for a collection
based request. At the end of processing, we take the 'count' param in
the request in to consideration. That may be just 10, and we have
processed a couple of thousands of items unnecessarily. Then, after
the page collection articles have been assembled, we re-iterate over
them to filter out the present/missing depending on the case.

The patch improves the performance for page collection recommendation
fetching by:
1. Stopping the processing when the desired number of recommendations
   have been reached.
2. Checking if recommendation is missing/present in target language,
   at the same loop that it checks about the source language.

To achieve this, the patch removes the notion of finder_map, and calls
directly the methods to get the recommendations.

Bug: T381366
Change-Id: I35f8f94f48369919a550f1b90df7352d68d5539b