Fix source and target for event logging of page-collections endpoint
The "log_api_request" method that logs the analytics events for the
recommendation api, expects a source and a target language to be set.
However, the "/page-collections" endpoint, doesn't contain a source
or target language, and leaving them empty leads to this error:
https://phabricator.wikimedia.org/P71680

This patch fixes this issue by explicitly setting these values to
None.

I think that the schema should also be updated to take into account
other information except for seed, searchAlgorithm, source and target
languages, that it currently supports. Examples of the new data we
should log, is whether the request is for page suggestions, section
suggestions or page-collection information, the given topic, the
rank method, etc.

Change-Id: I806499594e1d57c49e12e09edef9014b8dbdb825