Consider special language codes while checking for article existence
The langlinks from MW api gives language codes that may not
match with domain. For example, in the case of Norwegian Bokmål,
domain is 'no', but langlinks will list it as 'nb'.
This will cause suggesting existing articles to create
in recommendation system.

As a fix, consider the domain-language code mapping
while checking for article existence. Add unit tests too

Bug: T306508
Change-Id: I786b076ea9b63ef3040be0802e480e85805197b3