Popular/search recommander: use domain code in lllang parameter
For special language codes, those where the language code and
subdomain code are not the same, the lllang parameter of
the query API is expecting the domain code. When the language code
is provided, no langlinks is returned and it is assumed that the
article results do not exist in the target language. This leads
to 1) article translation recommendations for article that already
exist and 2) no section recommendations at all.

Bug: T306508
Change-Id: I5a1f2590d3dd49d90fe3eed3ba9beee6a1ee35f4