Fully encode URLs for GET requests
In fetcher/get, URL encoding is set to skip ':', '|' and '+'.
This means that those characters are left intact and that '+'
is assumed to be a space by the receiving party. Therefore,
'Wiki99/LGBT+' is interpreted as 'Wiki99/LGBT ' and the
page collection is not found by the 'pagecollectionmetadata'
API endpoint. Which leads to a general failure of the page
collections update progress.

This patch let '+' be encoded in the get function prior to
sending the request. It also removes early encoding from
space to '+' in the build_search_query function since
encoding is a transport concern more than an application concern.

Bonus: this patch improves error handling around page collection
metadata to make troubleshooting similar issues easier in the future.

Bug: T395026
Change-Id: I698f99da6a50603a55b3bae9485f989ce09e050e