Support size filtering for collection with source_lang=en
During collection cache update process, fetch "en" article size
so it can be used for size filtering. When source language is
not "en", filtering is not available.

This is a compromise between cache update performance and filtering
capabilities. Since source language is "en" 83% of the time, I
think it's an acceptable one.

On my local system, cache update went from 1m59s to 2m30s with
this change.

Change-Id: I8ef267a0d2fc0168ade6324ffa3160f51fff0a72