Increase httpx client max_connections for better consurency
Concurrent requests are used throughout this codebase to
reduce latency. This is controlled in part by
`API_CONCURRENCY_LIMIT: int = 10` but the httpx client
has a max connections limit of 5.

On my local system, setting the max connections to 20 reduces
the cache update time from 1:59 to 1:05 and it stays around
1:00 even if I increase to 50 or 100 so 20 may be a good
compromise.

Change-Id: Iac2ece8f9fe1f2799031d973c05c44ddeb9f247c