Don't treat CX Server 404 response as error
CX Server returns HTTP 404 (not found) to indicate that page
does not exist in the target language. This is simply
informational, it's not an error. However, the fetcher's get and
post functions raise an error in this case and that gets logged
in logstash, resulting in a lot of noise.

This patch introduces an option to accept HTTP 404 as success
for calls to CX Server.

Change-Id: I1787c9b3ee00716c4ee0c6ce7d915403ce9d3fe7