Downgrade unknown prefix from error to debug
Page collections sometimes contain interwiki links to sites
other than wikipedia. Their prefix include xtools and toollabs.

These links are not relevant in the context of translation
suggestions and are be ignored. However, they used to be logged
as error therefore adding noise to the logs.

This patch downgrades to debug level.

Bug: T404976
Change-Id: I6c6929cc2ad78843843e3a469b5cfde71e1223ed