Exclude collections with no articles from cache
Page collections cache includes collections with no articles
and untagged collections, likely due to cache polution when
two collections are tagged with the same name.

This patch includes the collection name in the cache
key to prevent matching and force updating when the name
changes.

It add additional checks to prevent caching of empty collections.

Bug: T407895
Bug: T407894
Change-Id: I8dcbfba60389f3d40f268edceeea0a8f0eaea563