Improve periodic update flow and error handling
The periodic cache update is running in a simple `while True:`
loop. If anything causes this loop to stop, like a unhandled
exception or a return statement, then the current instance is
left with an outdated or empty page collections cache.

This patch removes the return statements after errors with
updating the interwiki map and sitematrix.

It also introduces easier to find messages.

Bug: T406854
Change-Id: Ie832468dd3079c0bd9ad34687b4d7d49d5ed8b87