page collection groups: include single ‘/’ collections without siblings
Currently, when a page collection name contains a "/", but no other
collections sharing its prefix, that collection is omitted from the
"/page-collection-groups" API response.

For example, if there is a collection named "Tamil Wikipedia/Medical"
but no other collection starting with "Tamil Wikipedia/", the
collection is not included in the API response.

This patch fixes this issue by adding such collections to the
"ungrouped" list.

Change-Id: Iba795098efdb59fe3af422f82a0f87bb2e32dc20