collection recs: fix lead section size filtering
Currently, no results are returned when "lead_section" URL param is set,
and min/max sizes correspond to section sizes (not article sizes). This
happens because the recommendations are filtered based on their article
size, with the given min/max section sizes which are too low.

This patch fixes this issue by checking if article size filtering
should be applied. Another issue fixed by this patch is the fact that
until now, "lead_section_size" is added to the payload of the
recommendations, even if the "lead_section" URL param is missing,
which is not the intended behaviour.

Example URL demonstrating this undesired behaviour:
https://api.wikimedia.org/service/lw/recommendation/api/v1/translation?source=en&target=el&collections=true

Implementation note: A new "filter_recommendations_by_lead_section_size"
helper function has been added, with a clearer implementation. This
function can be used in other recommenders too, to avoid the code
duplication that we had until now.

Bug: T403730
Change-Id: I5366dec8e4b693044835a611cc09eabf86012275