add support for pagination for single page collections
This adds support for an alternative to classic pagination, with the
newly added "continue_offset" and "continue_seed" parameters.

When a single collection is requested, and "continue_offset" parameter
is set, we skip all the collection articles before the given index in
that parameter.

The "continue_seed" parameter on the other hand, provides the option
to provide an integer, in order to be able to always get the same
"randomly" shuffled recommendations.

This feature will allow us to have pagination for a single collection,
without being stuck always to the same "predefined" order.

Examples:
Start from the beginning and fetch 2 first section suggestions:

http://localhost:8080/api/v1/translation/sections?source=en&target=ig&count=2&collections=true&seed=Wiki99/food&continue_offset=

Then use the "continue_offset" and the "continue_seed" given in the above
API response, to get the next two recommendations:
http://localhost:8080/api/v1/translation/sections?source=en&target=ig&count=2&collections=true&seed=Wiki99/food&continue_offset=5&continue_seed=2343792424

Bug: T384485
Change-Id: I2d53d1208d60cf1e3c46189324d02355143b072e