Page collections caching: Use sitematrix lang code for all articles
Currently, when wikidata article titles are grouped by their languages,
the language used is just their dbnames without the "wiki" postfix.
That's works well for regular language codes (e.g. "enwiki" -> "en",
"eswiki" -> "es"), but not for special codes (e.g. for "zh_yuewiki"
we get "zh_yue", but the actual language code is "yue").

On the other hand, for wikipedia interwiki links, the "prefix" property
of the link is used, which for links for "zh-yue" Wikipedia, it is
"zh-yue", instead of "yue", which is the actual language code.

This discrepancy in the language codes, leads to duplicate wikidata
articles being leaked in page collection cache.

This patch fixes this issue, by introducing to new helper functions
that return the sitematrix language, when given a prefix or a dbname.

Bug: T410387
Change-Id: Iaa12935e344b05327333f704d960a53851a1b8ab