fix lead section size filtering for single page collections
For single page collections, the returned recommendations should follow
a specific order, to properly support pagination. However, when lead
section size filtering is applied, that order is lost leading to broken
pagination.

This patch fixes this issue, by keeping the original order, even when
lead section size filtering is applied. To achieve this, a new method
that supports both concurrent fetching and preserved order is added.
This method is slower than the "unordered" fetching, and thus it
should be used, only when "ordered" fetching is required.

Bug: T384485
Change-Id: I631c9991d4234d158b7eef384a67ae19a5656ab8