add support for combining single page collection with topic filter
This patch adds a new "featured_collection" parameter, that can be
combined with the "topic" filter to get (up to) 50% of the results
from the intersection of featured_collection+topic filters, and 50%
of regular random topic-based suggestions. The percentage (50%) is
controlled by a "ratio" parameter that defaults to 0.5.

To achieve this, the patch fetches the pageids of all the pages in
a page collection and uses the "pageid" search keyword to filter
topic-based recommendations just from these pages.

Example usage:
http://localhost:8080/api/v1/translation?source=en&target=el&topic=biography&featured_collection=African%20Film%20%26%20Cinema%20Topic%20list%202025&count=20

Bug: T409338
Change-Id: I3a435ef0545740797d1b8a2c4f4e1a6e9a8c26a4